Fix tab-detaching functionality
The problem, as explained in the comment -- if we try to create a window
with the currently-selected tab as the new first tab ID, that window
will be focused, and (as best I can tell), we won't have access to the
old tabs anymore (or, at least, we can't easily move them to the new
window anymore). To get around this, we do a check whether the first of
the tabs-to-move is active. If it is, we start by moving the last tab.